api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvd.studio>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
fastapi = { extras = ["standard"], version = "^0.115.5" }
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
langchain = "^0.3.7"
langchain-ollama = "^0.2.0"
langchain-core = "^0.3.18"
unstructured = { extras = ["image"], version = "^0.16.5" }
langchain-community = "^0.3.7"
langchain-unstructured = "^0.1.5"
pillow = "^11.0.0"
pillow-heif = "^0.20.0"
azure-ai-documentintelligence = "^1.0.0b4"
python-magic = "^0.4.27"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.3"
ipython = "^8.29.0"
jupyter = "^1.1.1"
notebook = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C", "B", "UP"]
ignore = ["E203", "E266", "E501", "E402"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query"]

[tool.ruff.lint.flake8-builtins]
builtins-ignorelist = ["id"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["pydantic.validator"]
